// apps/bitagent-launch-kernel/src/types.ts
export type IntegrationErrorCode =
  | "quote_error"
  | "router_submit_error"
  | "swap_timeout"
  | "egress_not_found"
  | "utxo_parse_error"
  | "utxo_ref_map_error"
  | "tradelayer_build_error"
  | "tradelayer_submit_error"
  | "wallet_sync_error"
  | "dlc_prepare_error"
  | "tradelayer_trade_error"
  | "filecoin_quote_error"
  | "filecoin_deal_error"
  | "compute_quote_error"
  | "compute_lease_error"
  | "chain_abstraction_error"
  | "near_intents_quote_error"
  | "near_intents_status_error"
  | "near_chain_signature_error"
  | "settlement_error"
  | "ledger_error"
  | "signer_broker_error"
  | "market_risk_error"
  | "recovery_error";
// ...
export class IntegrationBoundaryError extends Error {
  readonly code: IntegrationErrorCode;
  readonly causeData?: unknown;

  constructor(code: IntegrationErrorCode, message: string, causeData?: unknown) {
    super(message);
    this.name = "IntegrationBoundaryError";
    this.code = code;
    this.causeData = causeData;
  }
}